Some bug fixes, enhancements
- Updates unit test runner scripts
- More consistent and more precise computation of rounded surfaces
- Bugfix with :bias spec

+++ tests/run_tests.sh
@@ -2,9 +2,26 @@
 
 export KLAYOUT_HOME=/dev/null
 
-echo "Using KLayout:"
-klayout -v 
-echo ""
+def_path_win="$HOME/AppData/Roaming/KLayout" 
+if [ -e "$def_path_win" ] && [ -d "$def_path_win" ]; then
+  echo "Adding default installation path: $def_path_win"
+  export PATH="$def_path_win:$PATH"
+fi
+
+# locate klayout binary
+for exe in klayout klayout_app.exe; do
+  if which $exe 2>/dev/null; then
+    klayout=$(which $exe)
+    break
+  fi
+done
+
+if [ "$klayout" = "" ]; then
+  echo "No KLayout binary found in PATH .. please set path to point to klayout binary"
+  exit 1
+fi
+
+echo "Using KLayout binary from: $klayout"
 
 rm -rf run_dir
 mkdir -p run_dir
@@ -36,9 +53,9 @@ for tc_file in $tc_files; do
     xs_cut="-1,0;1,0"
   fi
 
-  klayout -rx -z -rd xs_run=$tc.xs -rd xs_cut="$xs_cut" -rd xs_out=run_dir/$tc.gds "$xs_input" -r $bin 
+  $klayout -rx -z -rd xs_run=$tc.xs -rd xs_cut="$xs_cut" -rd xs_out=run_dir/$tc.gds "$xs_input" -r $bin 
 
-  if klayout -b -rd a=au/$tc.gds -rd b=run_dir/$tc.gds -rd tol=10 -r run_xor.rb; then
+  if $klayout -b -rd a=au/$tc.gds -rd b=run_dir/$tc.gds -rd tol=10 -r run_xor.rb; then
     echo "No differences found."
   else
     failed="$failed $tc"
